+++ Tile2D/GUI/Canvas/Drawable/Color.cpp
@@ -16,6 +16,6 @@
 
 #include "Color.h"
 
-Color::Color(double red, double green, double blue) : red(red), green(green), blue(blue) {
+Color::Color(float red, float green, float blue) : red(red), green(green), blue(blue) {
 
 }
