+++ Tile2D/Physics/MapCollisionEventArgs.cpp
@@ -16,7 +16,7 @@
 
 #include "MapCollisionEventArgs.h"
 
-MapCollisionEventArgs::MapCollisionEventArgs(double deltaTime, Vec contactNormal, Vec tileCoordinates, Tile *tile) {
+MapCollisionEventArgs::MapCollisionEventArgs(float deltaTime, Vec contactNormal, Vec tileCoordinates, Tile *tile) {
     this->deltaTime = deltaTime;
     this->contactNormal = contactNormal;
     this->tileCoordinates = tileCoordinates;
